Add ruff and mypy gates, and fix the 167 findings they surfaced
The project had no linter, formatter, or type checker, and CI never ran the
eslint config the frontend already shipped. The codebase is heavily annotated,
but with nothing enforcing those annotations they were decoration rather than
a guarantee — so this wires both up and fixes everything they found.

Most of the 89 ruff and 78 mypy findings were mechanical, but three were real:

- Eight zip() calls pair closes[] with times[] or grid keys with grid combos.
  Both are equal-length by construction, so a length mismatch means corrupt
  bar data — exactly the bug a backtester must not swallow. strict=True now
  raises instead of silently truncating to the shorter series.
- Order.symbol/side/submitted_at are Optional on the SDK model, and the code
  dereferenced them unguarded: one malformed order would take down the whole
  trading cycle. Such orders are now skipped with a warning.
- _write_candidate and write_tuned_params could emit an evidence block full of
  nulls, which later reads as fact next to live trading params. They now
  refuse rather than write unfalsifiable evidence.

Most mypy noise came from alpaca-py typing every return as Union[Model, dict]
for its raw_data mode. Narrowing that once per call with cast() at the broker
boundary, rather than scattering type: ignore, keeps everything downstream
genuinely checked. That paid for itself immediately: mypy caught a rename in
research/cli.py that had been applied to five of six references.

B008 is suppressed for SimParams only, which is a frozen dataclass — the rule
targets mutable defaults, and sharing one immutable instance is the intent.

Dependencies are pinned with a lockstep requirements-dev.txt, so a green CI
run means something reproducible rather than "whatever released this morning".

+++ bot/config.py
@@ -138,7 +138,7 @@ def clamp_tunables(params: dict) -> dict:
             continue
         value = max(lo, min(hi, float(params[key])))
         if isinstance(getattr(Settings, key), int):
-            value = int(round(value))
+            value = round(value)
         out[key] = value
     # A near-equal EMA pair would make the trend filter meaningless;
     # require the slow period to exceed the fast by more than 3.
